+++ lib/helpers/extensions.dart
@@ -144,6 +144,34 @@ extension NyDateTime on DateTime? {
     return this.isInPast();
   }
 
+  /// Check if the date is in the morning.
+  bool isMorning() {
+    if (this == null) return false;
+    if (this?.hour == null) return false;
+    return this!.hour >= 0 && this!.hour < 12;
+  }
+
+  /// Check if the date is in the afternoon.
+  bool isAfternoon() {
+    if (this == null) return false;
+    if (this?.hour == null) return false;
+    return this!.hour >= 12 && this!.hour < 18;
+  }
+
+  /// Check if the date is in the evening.
+  bool isEvening() {
+    if (this == null) return false;
+    if (this?.hour == null) return false;
+    return this!.hour >= 18 && this!.hour < 24;
+  }
+
+  /// Check if the date is in the night.
+  bool isNight() {
+    if (this == null) return false;
+    if (this?.hour == null) return false;
+    return this!.hour >= 0 && this!.hour < 6;
+  }
+
   /// Format [DateTime] to DateTimeString - yyyy-MM-dd HH:mm:ss
   String? toDateTimeString() {
     if (this == null) return null;
@@ -1118,13 +1146,35 @@ extension NyBackpack<T> on String {
 
   /// Read a StorageKey value from NyStorage
   Future<T?> read<T>({dynamic defaultValue}) async {
-    return await fromStorage(defaultValue: defaultValue);
+    return await fromStorage<T>(defaultValue: defaultValue);
+  }
+
+  /// Read a JSON value from NyStorage
+  Future<T?> readJson<T>({dynamic defaultValue}) async {
+    T? response = await NyStorage.readJson<T>(this, defaultValue: defaultValue);
+    if (response == null) return null;
+    try {
+      return response;
+    } catch (e) {
+      NyLogger.error(e);
+      return null;
+    }
   }
 
   /// Store a value in NyStorage
   /// You can also store a value in the backpack by setting [inBackpack] to true
   store(dynamic value, {bool inBackpack = false}) async {
     return await NyStorage.store(this, value, inBackpack: inBackpack);
+  }
+
+  /// Store a JSON value in NyStorage
+  /// You can also store a value in the backpack by setting [inBackpack] to true
+  storeJson(dynamic value, {bool inBackpack = false}) async {
+    try {
+      return await NyStorage.storeJson(this, value, inBackpack: inBackpack);
+    } catch (e) {
+      NyLogger.error(e);
+    }
   }
 
   /// Add a value to a collection in NyStorage